solved 2

+++ src/main/java/school/mjc/stage0/module3/task4/SettingTypes.java
@@ -6,7 +6,7 @@ public class SettingTypes {
         char second = 'a';
         boolean third = 5 > 9;
         double forth = 12.90434323324342;
-        double fifth = 12.3;
+        float fifth = 12.3F;
         byte sixth = 1;
         short seventh = 129;
     }
